chore: Removed deprecated VobHelper.
